+++ Frends.SFTP.DownloadFiles/Frends.SFTP.DownloadFiles/Definitions/TransferState.cs
@@ -17,6 +17,7 @@ internal enum TransferState
     CleanUpFiles,
     CheckIfDestinationFileExists,
     RestoreModified,
-    Connection
+    Connection,
+    CheckSourceFiles
 }