+++ Frends.SFTP.DownloadFiles/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## [2.3.0]
+## [2.4.0] -2022-11-03
+### Added
+- [Breaking] Added parameters for the file extension of temporary source and destination files when rename options are enabled.
+- Fixed task to enable transfers with file paths. 
+- Fixed operations log to show correct state when source files are not found with filePaths.
+- Fixed operations log to use temp work path when getting source files to temp directory.
+- Added tests for the filePaths.
+- Modified tests to not use physical files but to create them during test run.
+
+## [2.3.0] - 2022-10-12
 ### Added
 - Added boolean parameter for adding a new line when appending to an existing file.
 - Changed the appending to use AppendAllText instead of AppendAllLines.